Changed the old [160][240] array to a single [38400] to we can move pixels around; Adds logs support; Adds support to select both background and foreground colors in console package

+++ cmd/mode3-animation/main.go
@@ -44,7 +44,7 @@ func drawImageMode3(palette map[uint16]uint8, image []uint16, width, height, ima
 	for y = 0; y < height; y++ {
 		for x = 0; x < width; x++ {
 			value := image[(x+imageWidth*y)+(offsetX+offsetY)]
-			mode3.SetPixel(uint8(x), uint8(y), value)
+			mode3.SetPixel(uint16(x), uint16(y), value)
 		}
 	}
 }
